% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiet.R
\name{quiet}
\alias{quiet}
\title{quiet function to suppresses pesky MannKendall warnings}
\usage{
quiet(x)
}
\arguments{
\item{x}{MannKendall stat run?}
}
\value{
\itemize{
\item A quiet run of the MannKendall function
}
}
\description{
via Hadley Wickham This function silences MannKendall function warnings
}
