% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_pearson_pair.R
\name{summarise_pearson}
\alias{summarise_pearson}
\title{Run and summarize Pearson correlation analysis}
\usage{
summarise_pearson(df_in, grouping_var_in, var1_A, var1_B)
}
\arguments{
\item{df_in}{-input dataframe with IDs as rownames, 1 grouping column, and at least two columns of data to correlate}

\item{grouping_var_in}{-quoted name of grouping variable}

\item{var1_A}{quoted names of variables for pairwise comparison}

\item{var1_B}{quoted names of variables for pairwise comparison}
}
\value{
dfvar_final dataframe containing the following columns; corr_var1-	strings containing the names of the variables compared;  corr_var2- strings containing the names of the variables compared; grouping_var_name- name of grouping variable; grouping_var- value of grouping variable for pairwise comparison; pearson_val- correlation coefficient value aka estimate;  p_val-  p-value;  data_n- number of data points included in comparison
}
\description{
Runs correlation cor.test and provides long format output with correlation
coefficient, pval, and sample size.
Rerun function for various correlation pairs and use bind_rows to combine dataframes
}
